requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
dj-database-url==2.3.0
Django==4.2.15
django-tailwind==3.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pygments==2.19.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
